+++ schnorr_fun/src/lib.rs
@@ -43,8 +43,6 @@ mod libsecp_compat;
 #[doc(hidden)]
 macro_rules! test_instance {
     () => {
-        $crate::Schnorr::<sha2::Sha256, _>::new(
-            $crate::nonce::Deterministic::<sha2::Sha256>::default(),
-        )
+        $crate::Schnorr::<sha2::Sha256, secp256kfun::nonce::Deterministic<sha2::Sha256>>::default()
     };
 }